edit #1
README 작성
폴더구조 개선
디자인 구조 개선
