+++ app/lib/definitions/types.ts
@@ -1,10 +1,5 @@
-// useInterval
-export interface CallbackFunction {
-  (): void;
-}
-
-// fetch lotto data
-export interface LottoData {
+// Lotto data fetched with fetch
+export type LottoData = {
   drwNo: number;
   drwtNo1: number;
   drwtNo2: number;
@@ -17,8 +12,9 @@ export interface LottoData {
   firstAccumamnt: number;
   firstWinamnt: number;
   firstPrzwnerCo: number;
-}
+};
 
+// lotto reducer action
 export type LottoAction =
   | {
       type: 'EXTRACTION';
@@ -36,23 +32,22 @@ export type LottoAction =
       suggestion: number[][];
     };
 
-export type InitialState = {
-  lottos: number[][];
-  suggestion: number[][];
-};
-
-export type Lottos = {
-  lottos: number[][];
-};
-
-export type Suggestion = {
-  suggestion: number[][];
-};
-
+// lotto context
 export type LottoContextValue = {
   state: {
     lottos: number[][];
     suggestion: number[][];
   };
   dispatch: React.Dispatch<LottoAction>;
+};
+
+// extractionLogic.ts numStorage value
+export type LottoNumber = {
+  sections: number[][];
+};
+
+// balls.ts lotto range
+export type Part = {
+  range: number[];
+  name: string;
 };
