+++ app/ui/lotto/lottoNumGenerator/lottoNumGenerator.module.scss
