+++ app/ui/lotto/showGeneratedLotto/paging.module.scss
