+++ app/ui/lotto/suggestionLotto/suggestionLotto.module.scss
